+++ oshi-core/src/main/java/oshi/hardware/platform/unix/freebsd/FreeBsdCentralProcessor.java
@@ -34,10 +34,11 @@ import org.slf4j.LoggerFactory;
 import com.sun.jna.Memory; // NOSONAR squid:S1191
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
-import com.sun.jna.ptr.IntByReference;
+import com.sun.jna.platform.unix.LibCAPI.size_t;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.hardware.common.AbstractCentralProcessor;
+import oshi.jna.platform.unix.NativeSizeTByReference;
 import oshi.jna.platform.unix.freebsd.FreeBsdLibc;
 import oshi.jna.platform.unix.freebsd.FreeBsdLibc.CpTime;
 import oshi.util.ExecutingCommand;
@@ -270,7 +271,8 @@ final class FreeBsdCentralProcessor extends AbstractCentralProcessor {
         Pointer p = new Memory(arraySize);
         String name = "kern.cp_times";
         // Fetch
-        if (0 != FreeBsdLibc.INSTANCE.sysctlbyname(name, p, new IntByReference((int) arraySize), null, 0)) {
+        if (0 != FreeBsdLibc.INSTANCE.sysctlbyname(name, p, new NativeSizeTByReference(new size_t(arraySize)), null,
+                size_t.ZERO)) {
             LOG.error("Failed syctl call: {}, Error code: {}", name, Native.getLastError());
             return ticks;
         }
@@ -315,9 +317,9 @@ final class FreeBsdCentralProcessor extends AbstractCentralProcessor {
     @Override
     public long queryContextSwitches() {
         String name = "vm.stats.sys.v_swtch";
-        IntByReference size = new IntByReference(FreeBsdLibc.INT_SIZE);
-        Pointer p = new Memory(size.getValue());
-        if (0 != FreeBsdLibc.INSTANCE.sysctlbyname(name, p, size, null, 0)) {
+        NativeSizeTByReference size = new NativeSizeTByReference(new size_t(FreeBsdLibc.INT_SIZE));
+        Pointer p = new Memory(size.getValue().longValue());
+        if (0 != FreeBsdLibc.INSTANCE.sysctlbyname(name, p, size, null, size_t.ZERO)) {
             return 0L;
         }
         return ParseUtil.unsignedIntToLong(p.getInt(0));
@@ -326,9 +328,9 @@ final class FreeBsdCentralProcessor extends AbstractCentralProcessor {
     @Override
     public long queryInterrupts() {
         String name = "vm.stats.sys.v_intr";
-        IntByReference size = new IntByReference(FreeBsdLibc.INT_SIZE);
-        Pointer p = new Memory(size.getValue());
-        if (0 != FreeBsdLibc.INSTANCE.sysctlbyname(name, p, size, null, 0)) {
+        NativeSizeTByReference size = new NativeSizeTByReference(new size_t(FreeBsdLibc.INT_SIZE));
+        Pointer p = new Memory(size.getValue().longValue());
+        if (0 != FreeBsdLibc.INSTANCE.sysctlbyname(name, p, size, null, size_t.ZERO)) {
             return 0L;
         }
         return ParseUtil.unsignedIntToLong(p.getInt(0));