+++ oshi-core/src/main/java/oshi/hardware/platform/unix/freebsd/FreeBsdSensors.java
@@ -25,10 +25,11 @@ package oshi.hardware.platform.unix.freebsd;
 
 import com.sun.jna.Memory; // NOSONAR squid:S1191
 import com.sun.jna.Pointer;
-import com.sun.jna.ptr.IntByReference;
+import com.sun.jna.platform.unix.LibCAPI.size_t;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.hardware.common.AbstractSensors;
+import oshi.jna.platform.unix.NativeSizeTByReference;
 import oshi.jna.platform.unix.freebsd.FreeBsdLibc;
 
 /**
@@ -50,11 +51,11 @@ final class FreeBsdSensors extends AbstractSensors {
      */
     private static double queryKldloadCoretemp() {
         String name = "dev.cpu.%d.temperature";
-        IntByReference size = new IntByReference(FreeBsdLibc.INT_SIZE);
-        Pointer p = new Memory(size.getValue());
+        NativeSizeTByReference size = new NativeSizeTByReference(new size_t(FreeBsdLibc.INT_SIZE));
+        Pointer p = new Memory(size.getValue().longValue());
         int cpu = 0;
         double sumTemp = 0d;
-        while (0 == FreeBsdLibc.INSTANCE.sysctlbyname(String.format(name, cpu), p, size, null, 0)) {
+        while (0 == FreeBsdLibc.INSTANCE.sysctlbyname(String.format(name, cpu), p, size, null, size_t.ZERO)) {
             sumTemp += p.getInt(0) / 10d - 273.15;
             cpu++;
         }