+++ oshi-core/src/main/java/oshi/util/platform/unix/openbsd/OpenBsdSysctlUtil.java
@@ -30,9 +30,10 @@ import com.sun.jna.Memory; // NOSONAR squid:S1191
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
 import com.sun.jna.Structure;
-import com.sun.jna.ptr.IntByReference;
+import com.sun.jna.platform.unix.LibCAPI.size_t;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.jna.platform.unix.NativeSizeTByReference;
 import oshi.jna.platform.unix.openbsd.OpenBsdLibc;
 import oshi.util.ExecutingCommand;
 import oshi.util.ParseUtil;
@@ -42,6 +43,8 @@ import oshi.util.ParseUtil;
  */
 @ThreadSafe
 public final class OpenBsdSysctlUtil {
+
+    private static final String SYSCTL_N = "sysctl -n ";
 
     private static final Logger LOG = LoggerFactory.getLogger(OpenBsdSysctlUtil.class);
 
@@ -60,9 +63,9 @@ public final class OpenBsdSysctlUtil {
      * @return The int result of the call if successful; the default otherwise
      */
     public static int sysctl(int[] name, int def) {
-        IntByReference size = new IntByReference(OpenBsdLibc.INT_SIZE);
-        Pointer p = new Memory(size.getValue());
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, p, size, null, 0)) {
+        NativeSizeTByReference size = new NativeSizeTByReference(new size_t(OpenBsdLibc.INT_SIZE));
+        Pointer p = new Memory(size.getValue().longValue());
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, p, size, null, size_t.ZERO)) {
             LOG.error(SYSCTL_FAIL, name, Native.getLastError());
             return def;
         }
@@ -79,9 +82,9 @@ public final class OpenBsdSysctlUtil {
      * @return The long result of the call if successful; the default otherwise
      */
     public static long sysctl(int[] name, long def) {
-        IntByReference size = new IntByReference(OpenBsdLibc.UINT64_SIZE);
-        Pointer p = new Memory(size.getValue());
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, p, size, null, 0)) {
+        NativeSizeTByReference size = new NativeSizeTByReference(new size_t(OpenBsdLibc.UINT64_SIZE));
+        Pointer p = new Memory(size.getValue().longValue());
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, p, size, null, size_t.ZERO)) {
             LOG.warn(SYSCTL_FAIL, name, Native.getLastError());
             return def;
         }
@@ -99,14 +102,14 @@ public final class OpenBsdSysctlUtil {
      */
     public static String sysctl(int[] name, String def) {
         // Call first time with null pointer to get value of size
-        IntByReference size = new IntByReference();
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, null, size, null, 0)) {
+        NativeSizeTByReference size = new NativeSizeTByReference();
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, null, size, null, size_t.ZERO)) {
             LOG.warn(SYSCTL_FAIL, name, Native.getLastError());
             return def;
         }
         // Add 1 to size for null terminated string
-        Pointer p = new Memory(size.getValue() + 1L);
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, p, size, null, 0)) {
+        Pointer p = new Memory(size.getValue().longValue() + 1L);
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, p, size, null, size_t.ZERO)) {
             LOG.warn(SYSCTL_FAIL, name, Native.getLastError());
             return def;
         }
@@ -123,8 +126,8 @@ public final class OpenBsdSysctlUtil {
      * @return True if structure is successfuly populated, false otherwise
      */
     public static boolean sysctl(int[] name, Structure struct) {
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, struct.getPointer(), new IntByReference(struct.size()),
-                null, 0)) {
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, struct.getPointer(),
+                new NativeSizeTByReference(new size_t(struct.size())), null, size_t.ZERO)) {
             LOG.error(SYSCTL_FAIL, name, Native.getLastError());
             return false;
         }
@@ -141,13 +144,13 @@ public final class OpenBsdSysctlUtil {
      *         otherwise. Its value on failure is undefined.
      */
     public static Memory sysctl(int[] name) {
-        IntByReference size = new IntByReference();
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, null, size, null, 0)) {
+        NativeSizeTByReference size = new NativeSizeTByReference();
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, null, size, null, size_t.ZERO)) {
             LOG.error(SYSCTL_FAIL, name, Native.getLastError());
             return null;
         }
-        Memory m = new Memory(size.getValue());
-        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, m, size, null, 0)) {
+        Memory m = new Memory(size.getValue().longValue());
+        if (0 != OpenBsdLibc.INSTANCE.sysctl(name, name.length, m, size, null, size_t.ZERO)) {
             LOG.error(SYSCTL_FAIL, name, Native.getLastError());
             return null;
         }
@@ -168,7 +171,7 @@ public final class OpenBsdSysctlUtil {
      * @return The int result of the call if successful; the default otherwise
      */
     public static int sysctl(String name, int def) {
-        return ParseUtil.parseIntOrDefault(ExecutingCommand.getFirstAnswer("sysctl -n " + name), def);
+        return ParseUtil.parseIntOrDefault(ExecutingCommand.getFirstAnswer(SYSCTL_N + name), def);
     }
 
     /**
@@ -181,7 +184,7 @@ public final class OpenBsdSysctlUtil {
      * @return The long result of the call if successful; the default otherwise
      */
     public static long sysctl(String name, long def) {
-        return ParseUtil.parseLongOrDefault(ExecutingCommand.getFirstAnswer("sysctl -n " + name), def);
+        return ParseUtil.parseLongOrDefault(ExecutingCommand.getFirstAnswer(SYSCTL_N + name), def);
     }
 
     /**
@@ -194,7 +197,7 @@ public final class OpenBsdSysctlUtil {
      * @return The String result of the call if successful; the default otherwise
      */
     public static String sysctl(String name, String def) {
-        String v = ExecutingCommand.getFirstAnswer("sysctl -n " + name);
+        String v = ExecutingCommand.getFirstAnswer(SYSCTL_N + name);
         if (null == v || v.isEmpty()) {
             return def;
         }